Abstracted the parser
